==> R/00_doit.R <==
# Run all scripts
# ------------------------------------------------------------------------------
source(file = "R/00_packages.R")
source(file = "R/01_load.R")
source(file = "R/02_clean.R")
source(file = "R/03_augment.R")
source(file = "R/04_glm_death_prediction.R")
source(file = "R/05_lm(sbp_vs_dbp).R")
source(file = "R/06_metastases_plots.R")
source(file = "R/07_PCA.R")
source(file = "R/08_keras_ANN.R")
