==> R/00_packages.R <==
#install packages required 

install.packages("tidyverse")
install.packages("xlsx")
install.packages("ggplot2")
install.packages("stringr")
install.packages("broom")
install.packages("purrr")
install.packages("caret")
install.packages("caretEnsemble")
install.packages("gridExtra")
install.packages("keras")
install.packages("tensorflow")
